--- tests/fixtures/HTSeq-FPKM-UQ.csv
@@ -1,4 +1,0 @@
-    TCGA-04-1331-01A	TCGA-04-1332-01A	TCGA-04-1338-01A	TCGA-04-1341-01A	TCGA-04-1343-01A	TCGA-04-1347-01A	TCGA-04-1350-01A	TCGA-04-1356-01A	TCGA-04-1357-01A	TCGA-04-1361-01A
-ENSG00000000003.13	19.13	18.43	18.17	18.33	18.14	19.52	19.12	18.83	18.19	19.09
-ENSG00000000005.5	11.14	12.17	9.344	10.84	15.29	12.24	0	8.896	10.59	13.12
-ENSG00000000419.11	19.91	19.34	20.21	20.05	21.22	21.28	20.27	20.11	19.78	19.46
