--- tests/fixtures/HTSeq-FPKM-UQ.json
@@ -1,15 +1,0 @@
-{"cohort":"GDC Pan-Cancer (PANCAN)",
- "url":"https://docs.gdc.cancer.gov/Data/Release_Notes/Data_Release_Notes/#data-release-90, https://api.gdc.cancer.gov/data/",
- "dataSubType":"gene expression RNAseq",
- ":probeMap":"gencode.v22.annotation.gene.probeMap",
- "dataProducer":"Genomic Data Commons",
- "label":"HTSeq - FPKM-UQ",
- "colNormalization":true,
- "wrangling_procedure":
- "Data from the same sample but from different vials/portions/analytes/aliquotes is averaged; data from different samples is combined into genomicMatrix; all data is then log2(x+1) transformed.",
- "PLATFORM":"Illumina",
- "type":"genomicMatrix",
- "wrangler":"Xena GDC ETL script",
- "version":"04-28-2019",
- "unit":"log2(fpkm-uq+1)"
-}
