--- tests/fixtures/df2.csv
@@ -1,1 +1,0 @@
-\t\tchrom\tchromstart\tvalue\n0\tTCGA-30-1891-01A\tchr1\t3301765\t0.1913\n1\tTCGA-61-2102-01A\tchr1\t3301765\t0.1153\n2\tTCGA-29-2429-01A\tchr1\t3301765\t-0.2114\n3\tTCGA-23-1023-01R\tchr1\t3301765\t-0.5136\n
