+++ tests/test_utils.py
@@ -1,11 +1,4 @@
-import json
 import os
-from mock import Mock, patch
-import time
-try:
-    from StringIO import StringIO
-except ImportError:
-    from io import StringIO
 
 import pandas as pd
 
@@ -20,28 +13,6 @@ def test_mkdir_p():
     assert os.path.isdir(input_) is True
     assert input_ in result
     os.rmdir(input_)
-
-
-@patch("sys.stdout", new_callable=StringIO)
-def test_equal_matrices(mocked_print):
-    utils.equal_matrices(PATH + "df1.csv", PATH + "df2.csv")
-    utils.equal_matrices(PATH + "df1.csv", PATH + "df3.csv")
-    assert mocked_print.getvalue() == "Not equal.\nEqual.\n"
-
-
-@patch.object(utils, 'time', Mock(wraps=time))
-def test_metadata():
-    utils.time.gmtime.return_value = time.struct_time(
-        (2019, 4, 28, 0, 0, 0, 0, 0, 0)
-    )
-    path_to_df = PATH + "HTSeq-FPKM-UQ.csv"
-    utils.metadata(path_to_df, "htseq_fpkm-uq")
-    with open(PATH + "HTSeq-FPKM-UQ.csv.json", "r") as actual:
-        actual = json.load(actual)
-    os.unlink(PATH + "HTSeq-FPKM-UQ.csv.json")
-    with open(PATH + "HTSeq-FPKM-UQ.json", "r") as expected:
-        expected = json.load(expected)
-    assert actual == expected
 
 
 def test_merge_xena():
