+++ tests/test_xena_dataset.py
@@ -5,6 +5,7 @@ from xena_gdc_etl import xena_dataset
 
 PATH = "tests/fixtures/xena_dataset/"
 
+
 def test_read_biospecimen():
     file_name = PATH + "nationwidechildrens.org_biospecimen.TCGA-AR-A0TQ"
     expected = xena_dataset.read_biospecimen(file_name + ".xml")
