+++ xena_gdc_etl/constants.py
@@ -1,3 +1,40 @@
+# Map GDC project_id to Xena specific cohort name.
+GDC_XENA_COHORT = {
+        'TCGA-BRCA': 'GDC TCGA Breast Cancer (BRCA)',
+        'TCGA-LUAD': 'GDC TCGA Lung Adenocarcinoma (LUAD)',
+        'TCGA-UCEC': 'GDC TCGA Endometrioid Cancer (UCEC)',
+        'TCGA-LGG': 'GDC TCGA Lower Grade Glioma (LGG)',
+        'TCGA-HNSC': 'GDC TCGA Head and Neck Cancer (HNSC)',
+        'TCGA-PRAD': 'GDC TCGA Prostate Cancer (PRAD)',
+        'TCGA-LUSC': 'GDC TCGA Lung Squamous Cell Carcinoma (LUSC)',
+        'TCGA-THCA': 'GDC TCGA Thyroid Cancer (THCA)',
+        'TCGA-SKCM': 'GDC TCGA Melanoma (SKCM)',
+        'TCGA-OV': 'GDC TCGA Ovarian Cancer (OV)',
+        'TCGA-STAD': 'GDC TCGA Stomach Cancer (STAD)',
+        'TCGA-COAD': 'GDC TCGA Colon Cancer (COAD)',
+        'TCGA-BLCA': 'GDC TCGA Bladder Cancer (BLCA)',
+        'TCGA-GBM': 'GDC TCGA Glioblastoma (GBM)',
+        'TCGA-LIHC': 'GDC TCGA Liver Cancer (LIHC)',
+        'TCGA-KIRC': 'GDC TCGA Kidney Clear Cell Carcinoma (KIRC)',
+        'TCGA-CESC': 'GDC TCGA Cervical Cancer (CESC)',
+        'TCGA-KIRP': 'GDC TCGA Kidney Papillary Cell Carcinoma (KIRP)',
+        'TCGA-SARC': 'GDC TCGA Sarcoma (SARC)',
+        'TCGA-ESCA': 'GDC TCGA Esophageal Cancer (ESCA)',
+        'TCGA-PAAD': 'GDC TCGA Pancreatic Cancer (PAAD)',
+        'TCGA-PCPG': 'GDC TCGA Pheochromocytoma & Paraganglioma (PCPG)',
+        'TCGA-READ': 'GDC TCGA Rectal Cancer (READ)',
+        'TCGA-TGCT': 'GDC TCGA Testicular Cancer (TGCT)',
+        'TCGA-LAML': 'GDC TCGA Acute Myeloid Leukemia (LAML)',
+        'TCGA-THYM': 'GDC TCGA Thymoma (THYM)',
+        'TCGA-ACC': 'GDC TCGA Adrenocortical Cancer (ACC)',
+        'TCGA-MESO': 'GDC TCGA Mesothelioma (MESO)',
+        'TCGA-UVM': 'GDC TCGA Ocular melanomas (UVM)',
+        'TCGA-KICH': 'GDC TCGA Kidney Chromophobe (KICH)',
+        'TCGA-UCS': 'GDC TCGA Uterine Carcinosarcoma (UCS)',
+        'TCGA-CHOL': 'GDC TCGA Bile Duct Cancer (CHOL)',
+        'TCGA-DLBC': 'GDC TCGA Large B-cell Lymphoma (DLBC)'
+    }
+
 # Map xena_dtype to corresponding metadata template.
 METADATA_TEMPLATE = {
     'htseq_counts': 'template.rna.meta.json',
@@ -16,6 +53,7 @@ METADATA_TEMPLATE = {
     'methylation27': 'template.methylation.meta.json',
     'methylation450': 'template.methylation.meta.json'
 }
+
 # Jinja2 template variables for corresponding "xena_dtype".
 METADATA_VARIABLES = {
     'htseq_counts': {'gdc_type': 'HTSeq - Counts'},
