+++ xena_gdc_etl/scripts/panTCGA.py
@@ -34,6 +34,44 @@ import pandas as pd
 from ..constants import METADATA_TEMPLATE, METADATA_VARIABLES  # noqa
 
 
+def metadata(matrix, xena_dtypes):
+    """Generating Xena metadata for a Xena matrix.
+
+    Args:
+        matrix (str): The path, including the filename, of the Xena matrix.
+        xena_dtypes (str): One data type code indication the data type in
+            matrices to be merged. Supported data type codes include (without
+            quotation marks): "htseq_counts", "htseq_fpkm", "htseq_fpkm-uq",
+            "mirna", "masked_cnv", "muse_snv", "mutect2_snv",
+            "somaticsniper_snv", "varscan2_snv", "GDC_phenotype", "survival",
+            "methylation27".
+    """
+
+    # Generate metadata.
+    print('Creating metadata file ...', end='')
+    sys.stdout.flush()
+    jinja2_env = jinja2.Environment(
+            loader=jinja2.PackageLoader('xena_gdc_etl', 'resources')
+        )
+    metadata_template = jinja2_env.get_template(METADATA_TEMPLATE[xena_dtypes])
+    matrix_date = time.strftime("%m-%d-%Y",
+                                time.gmtime(os.path.getmtime(matrix)))
+    variables = {
+            'project_id': 'GDC-PANCAN',
+            'date': matrix_date,
+            'gdc_release': 'https://docs.gdc.cancer.gov/Data/Release_Notes/Data_Release_Notes/#data-release-90', # noqa
+            'xena_cohort': 'GDC Pan-Cancer (PANCAN)'
+        }
+    try:
+        variables.update(METADATA_VARIABLES[xena_dtypes])
+    except KeyError:
+        pass
+    outmetadata = matrix + '.json'
+    with open(outmetadata, 'w') as f:
+        f.write(metadata_template.render(**variables))
+    print('\rMetadata JSON is saved at {}.'.format(outmetadata))
+
+
 def main():
     root_dir = r'/mnt/gdc/updates'
     out_dir = r'/mnt/gdc/updates/GDC-PANCAN/Xena_Matrices'
@@ -48,7 +86,7 @@ def main():
             k: os.path.join(meta_templates_dir, v)
             for k, v in METADATA_TEMPLATE.items()
         }
-    
+
     for dtype in datatypes:
         if dtype in ['htseq_counts', 'htseq_fpkm', 'htseq_fpkm-uq', 'mirna']:
             merge_axis = 1
@@ -80,7 +118,7 @@ def main():
         print('Saving merged matrix to {} ...'.format(outmatrix), end='')
         sys.stdout.flush()
         merged.to_csv(outmatrix, sep='\t', encoding='utf-8')
-        del merged # Prevent from doubling memory usage
+        del merged  # Prevent from doubling memory usage
         print('\rMerged "{}" matrix is ready at {}'.format(dtype, outmatrix))
         # Generate metadata.
         print('Creating metadata file ...', end='')
